Created router component for ModView
